Reworking init scripts, making bin less of the focus
