+++ module/base/bin/travetto-plugin-clean.js
@@ -1,0 +1,1 @@
+require('./lib').clean();
