+++ module/base/src/bootstrap/register.d.ts
@@ -1,0 +1,1 @@
+export function registerLoaders(): void;
