+++ module/base/src/scan-app.ts
@@ -70,12 +70,13 @@ export class ScanApp {
   }
 
   static getStandardAppFiles() {
-    return ScanApp
+    const res = ScanApp
       .findFiles('.ts', x =>
         !x.endsWith('.d.ts') && (
           /^(src\/|support\/|index)/.test(x) ||
           /(@travetto\/[^\/]+\/(src\/|support\/|index))/.test(x)
         ) && !x.includes('@travetto/test'))
       .map(x => x.file);
+    return res;
   }
 }
