+++ module/test/bin/test-worker.js
@@ -1,0 +1,5 @@
+// @ts-check
+const { prepareEnv, worker } = require(`${process.env.TRV_TEST_BASE || '..'}/bin/lib`);
+
+prepareEnv();
+worker();
