+++ module/test/src/worker/types.ts
@@ -1,5 +1,3 @@
-import { FsUtil } from '@travetto/base';
-
 export const Events = {
   RUN: 'run',
   RUN_COMPLETE: 'runComplete',
@@ -7,6 +5,3 @@ export const Events = {
   INIT_COMPLETE: 'initComplete',
   READY: 'ready'
 };
-
-// Handle bad symlink behavior, by allowing specifying full path.  Used for dev generally
-export const TEST_BASE = process.env.TRV_TEST_BASE || FsUtil.resolveUnix(__dirname, '../..');
